refactor: use os hook (#427)
